refactor(view): split keymaps_tool into schema, session, ai and results
keymaps_tool.lua was the largest and least coherent of the eleven keymap
sections at 445 lines, and its boundaries were inherited verbatim from the
old monolith rather than chosen: it held table diff, DDL, the schema
sidebar, two pickers, watch and write mode, open-as-editable, query
history, both AI entry points and the result registry.

Four groups fall out along the lines of what the key actually touches:
keymaps_schema (the table's structure), keymaps_session (what this grid is
bound to and how it reads/writes it), keymaps_ai, and keymaps_results (the
only two keys that iterate every other open grid via ctx.each_session).
The groups are contiguous runs of the old file, so concatenating the four
new bodies reproduces the old one byte for byte and the KEYMAP_SECTIONS
entries slot in where the single entry was. Registration order therefore
does not move, and it is the precedence rule for a duplicated lhs.

Verified with a dump of the ordered vim.keymap.set() sequence and of the
resolved buffer keymaps (104 in n, 11 in x): identical before and after,
apart from the informative definition-site column.
